requirements.txt
genshin>=1.7.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
PySocks>=1.7.1
Pillow>=10.0.0
pycryptodome>=3.20.0
qrcode>=7.4.2
